Pipeline msconvert & mzmine3 (#2)
## Summary
First draft of the pipeline with a driver script!

The main steps are:
1) Use msConvert to turn SCIEX raw output files into mzXML (run via singularity)
2) Prep the batch file for MZmine3
3) Run MZmine3 to process the data

And not yet included:
4) Process peak table with python script

I also updated the README with instructions to run the pipeline, as well as to install and set up msconvert and mzmine3.

## Test Plan
Tested on a couple test runs on the server. Not all of them ran fully but we think there are issues with some of the raw data in those.